api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via django
certifi==2021.5.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.6
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
defusedxml==0.7.1
    # via djangorestframework-xml
django==3.2.7
    # via
    #   -r requirements.in
    #   django-datetime-utc
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-yasg
django-datetime-utc==1.0.3
    # via -r requirements.in
django-debug-toolbar==3.2.2
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-csv
    #   djangorestframework-gis
    #   drf-amsterdam
    #   drf-extensions
    #   drf-yasg
djangorestframework-csv==2.1.1
    # via drf-amsterdam
djangorestframework-gis==0.17
    # via -r requirements.in
djangorestframework-xml==2.0.0
    # via
    #   -r requirements.in
    #   drf-amsterdam
drf-amsterdam==0.4.1
    # via -r requirements.in
drf-extensions==0.7.1
    # via drf-amsterdam
drf-yasg==1.20.0
    # via -r requirements.in
idna==3.2
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
markupsafe==2.0.1
    # via jinja2
packaging==21.0
    # via drf-yasg
psycopg2-binary==2.9.1
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.2
    # via django-datetime-utc
pytz==2021.3
    # via
    #   -r requirements.in
    #   django
requests==2.26.0
    # via
    #   -r requirements.in
    #   coreapi
ruamel.yaml==0.17.16
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sentry-sdk==1.4.3
    # via -r requirements.in
six==1.16.0
    # via
    #   djangorestframework-csv
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
unicodecsv==0.14.1
    # via djangorestframework-csv
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via
    #   requests
    #   sentry-sdk
